refactor(checkpoint): refactored checkpoint module into seperate files for cleaner debugging
