+++ src/verdant/panel/events/event-date.tsx
@@ -1,18 +1,22 @@
 import * as React from "react";
-import {connect} from "react-redux";
-import {verdantState} from "../../redux/index";
-import {Checkpoint, CheckpointType} from "../../../lilgit/model/checkpoint";
-import {dateOpen, dateClose, saveBundles, eventState} from "../../redux/events";
+import { connect } from "react-redux";
+import { verdantState } from "../../redux/index";
+import { Checkpoint, CheckpointType } from "../../../lilgit/model/checkpoint";
+import {
+  dateOpen,
+  dateClose,
+  saveBundles,
+  eventState,
+} from "../../redux/events";
 import NotebookEventDateBundle from "./event-date-bundle";
 
 /* CSS Constants */
-const DATE = "Verdant-events-date"
+const DATE = "Verdant-events-date";
 const DATE_HEADER = `${DATE}-header`;
 const DATE_HEADER_LABEL = `${DATE_HEADER}-label`;
 const DATE_HEADER_COLLAPSE = `${DATE_HEADER}-collapse`;
 const DATE_HEADER_COLLAPSE_COUNT = `${DATE_HEADER_COLLAPSE}-count`;
 const DATE_HEADER_COLLAPSE_ARROW = `${DATE_HEADER_COLLAPSE}-arrow`;
-
 
 const INTERVAL_WIDTH = 300000; // Max bundle time interval in milliseconds
 
@@ -37,29 +41,30 @@ class NotebookEventDate extends React.Component<NotebookDate_Props> {
   render() {
     return (
       <div className={DATE}>
-        <div className={DATE_HEADER}
-             onClick={() => {
-               if (this.props.isOpen) this.props.close(this.props.date_id);
-               else this.props.open(this.props.date_id);
-             }}>
+        <div
+          className={DATE_HEADER}
+          onClick={() => {
+            if (this.props.isOpen) this.props.close(this.props.date_id);
+            else this.props.open(this.props.date_id);
+          }}
+        >
           <div className={DATE_HEADER_LABEL}>
             {Checkpoint.formatDate(this.props.date)}
           </div>
           <div className={DATE_HEADER_COLLAPSE}>
             <div
               className={`${DATE_HEADER_COLLAPSE_COUNT} 
-              ${this.props.isOpen ? "hidden" : ""}`}>
+              ${this.props.isOpen ? "hidden" : ""}`}
+            >
               ({this.props.events.length})
             </div>
             <div
               className={`${DATE_HEADER_COLLAPSE_ARROW} 
-              ${this.props.isOpen ? "" : "closed"}`}>
-            </div>
+              ${this.props.isOpen ? "" : "closed"}`}
+            ></div>
           </div>
         </div>
-        <div>
-          {this.props.isOpen ? this.makeBundles() : null}
-        </div>
+        <div>{this.props.isOpen ? this.makeBundles() : null}</div>
       </div>
     );
   }
@@ -67,10 +72,14 @@ class NotebookEventDate extends React.Component<NotebookDate_Props> {
   private makeBundles() {
     /* Creates date bundles using bundled indices */
     let bundledIndices;
-    if (this.props.bundles === null || // There are no saved bundles
-      this.props.bundles.reduce( // The bundles are out of date
-        (a, x) => a + x.length, 0
-      ) !== this.props.events.length) {
+    if (
+      this.props.bundles === null || // There are no saved bundles
+      this.props.bundles.reduce(
+        // The bundles are out of date
+        (a, x) => a + x.length,
+        0
+      ) !== this.props.events.length
+    ) {
       // If there are no bundles or the bundles need an update, compute bundles
       bundledIndices = this.computeBundles(this.props.events);
       this.props.saveBundles(bundledIndices, this.props.date_id);
@@ -80,16 +89,14 @@ class NotebookEventDate extends React.Component<NotebookDate_Props> {
     }
 
     // Creates DateBundle for each set of dates
-    return bundledIndices.map(
-      (idx_list, i) => (
-          <NotebookEventDateBundle
-            key={i}
-            bundle_id={i}
-            events={[...idx_list]}
-            date_id={this.props.date_id}
-          />
-      )
-    )
+    return bundledIndices.map((idx_list, i) => (
+      <NotebookEventDateBundle
+        key={i}
+        bundle_id={i}
+        events={[...idx_list]}
+        date_id={this.props.date_id}
+      />
+    ));
   }
 
   private computeBundles(events: eventState[]): number[][] {
@@ -101,8 +108,8 @@ class NotebookEventDate extends React.Component<NotebookDate_Props> {
     return events.reduceRight(this.reducer, {
       accumulator: [],
       timeBound: Infinity,
-      lastType: null
-    }).accumulator
+      lastType: null,
+    }).accumulator;
   }
 
   private reducer(accObj: accumulatorObject, e: eventState, idx) {
@@ -111,22 +118,25 @@ class NotebookEventDate extends React.Component<NotebookDate_Props> {
     // Compute properties of current element
     let timeStamp = e.events[0].timestamp;
     let eventType = NotebookEventDate.eventType(e);
-    if ((timeStamp > accObj.timeBound) &&
-      (eventType === accObj.lastType)) { // add event to current bundle
-      const newAccumulator = accObj.accumulator.slice(0, -1).concat(
-        [accObj.accumulator[accObj.accumulator.length - 1].concat([idx])]
-      )
+    if (timeStamp > accObj.timeBound && eventType === accObj.lastType) {
+      // add event to current bundle
+      const newAccumulator = accObj.accumulator
+        .slice(0, -1)
+        .concat([
+          accObj.accumulator[accObj.accumulator.length - 1].concat([idx]),
+        ]);
       return {
         accumulator: newAccumulator,
         timeBound: accObj.timeBound,
-        lastType: accObj.lastType
-      }
-    } else { // create new bundle
+        lastType: accObj.lastType,
+      };
+    } else {
+      // create new bundle
       return {
         accumulator: accObj.accumulator.concat([[idx]]),
         timeBound: timeStamp - INTERVAL_WIDTH,
-        lastType: eventType
-      }
+        lastType: eventType,
+      };
     }
   }
 
@@ -134,9 +144,9 @@ class NotebookEventDate extends React.Component<NotebookDate_Props> {
     /* Helper for reducer.
        Returns CheckpointType if all checkpoints in event have same type,
        else returns null */
-    return e.events.map(c => c.checkpointType).reduce(
-      (acc, current) => acc === current ? acc : null
-    );
+    return e.events
+      .map((c) => c.checkpointType)
+      .reduce((acc, current) => (acc === current ? acc : null));
   }
 }
 
@@ -144,7 +154,7 @@ const mapDispatchToProps = (dispatch: any) => {
   return {
     open: (d) => dispatch(dateOpen(d)),
     close: (d) => dispatch(dateClose(d)),
-    saveBundles: (bundles, d) => dispatch(saveBundles(bundles, d))
+    saveBundles: (bundles, d) => dispatch(saveBundles(bundles, d)),
   };
 };
 
@@ -157,11 +167,8 @@ const mapStateToProps = (
     date: dateState.date,
     events: dateState.events,
     isOpen: dateState.isOpen,
-    bundles: dateState.bundles
+    bundles: dateState.bundles,
   };
 };
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(NotebookEventDate);
+export default connect(mapStateToProps, mapDispatchToProps)(NotebookEventDate);
